+++ membres/src/main/resources/schema.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS members;
 
 CREATE TABLE members (
-  id int AUTO_INCREMENT PRIMARY KEY NOT NULL,
+  id bigint AUTO_INCREMENT PRIMARY KEY NOT NULL,
   nom varchar(255) default NULL,
   mdp varchar(255),
   proposition varchar(255) default NULL,
